Prepare zuul job: bootstrap (#359)

+++ cleanup.sh
@@ -6,28 +6,20 @@ set -e
 # repository on a preinstalled Ubuntu and the osism/node image has
 # not served as the basis for it.
 
-systemctl stop apt-daily-upgrade.timer
-systemctl stop apt-daily.timer
-systemctl stop motd-news.timer
-systemctl stop multipathd.socket
-systemctl stop multipathd.service
-systemctl stop ufw.service
-systemctl stop unattended-upgrades.service
+for service in apt-daily-upgrade.timer apt-daily.timer motd-news.timer multipathd.socket multipathd.service ufw.service unattended-upgrades.service; do
+    if systemctl is-active --quiet multipathd.socket; then
+        systemctl stop multipathd.socket
+        systemctl disable multipathd.socket
+    fi
+done
 
-systemctl disable apt-daily-upgrade.timer
-systemctl disable apt-daily.timer
-systemctl disable motd-news.timer
-systemctl disable multipathd.socket
-systemctl disable multipathd.service
-systemctl disable ufw.service
-systemctl disable unattended-upgrades.service
-
-apt purge --yes snapd modemmanager lxd-agent-loader frr
-apt purge --yes plymouth plymouth-theme-ubuntu-text
-apt purge --yes ubuntu-advantage-tools xauth landscape-common btrfs-progs
-apt purge --yes apport apport-symptoms open-vm-tools ntfs-3g
-apt purge --yes telnet pastebinit tnftp ftp open-iscsi bolt packagekit
-apt autoremove --yes --purge
+apt-get update
+apt-get purge --yes snapd modemmanager lxd-agent-loader frr
+apt-get purge --yes plymouth plymouth-theme-ubuntu-text
+apt-get purge --yes ubuntu-advantage-tools xauth landscape-common btrfs-progs
+apt-get purge --yes apport apport-symptoms open-vm-tools ntfs-3g
+apt-get purge --yes telnet pastebinit tnftp ftp open-iscsi bolt packagekit
+apt-get autoremove --yes --purge
 
 if [[ -e /swap.img ]]; then
     swapoff /swap.img
